This check in incorporates the changes required to build with NumPy instead of Numeric. Building with Numeric is still the default. Defining NUMPY causes NUMPY to be built instead.
It also incorporates the capability to set file options in PyNio.
The constructor arguments are modified as follows:

 f = Nio.open_file(file_name,mode,options,history)

where all arguments except 'file_name' are optional. However, since the
arguments are position-dependent, to include any optional argument
requires that all preceding arguments be specified.

'options' is an instance of the Nio.options class.
It is instantiated as follows:

options = Nio.options()

Thereafter, in a similar fashion to Ngl.resources, key-value pairs
may be defined:

options.prefill = False
options.format = "largefile"

and so on.

Once the desired options are defined, it may be used as the third
argument to Nio.open_file.

Note that unlike Ncl where the options can be set globally, all
Nio options must be set for each file as it is opened, and are then fixed
until the file is closed.

Three options have different defaults from NCL:

NetCdf option defineMode defaults to True
NetCdf option suppressClose defaults to True
Grib option initialTimeCoordinateType defaults to "numeric"