Nio.py: added __array_module__ and __array_module_version__ as module attributes of Nio.
Fixed some problems with reading, creating and assigning to character array
variables. Now in the numpy version, you can specify either 'S1' or 'c' as
the typecode to create a character array variable.

Note that "string" arrays can be read in and displayed (the only current
example is the ensembleX_info variables that are creating for ensemble
GRIB files), but you cannot write variables of this type to a writable file.
Part of the problem is that with NetCDF (or with our HDF 4 interface)
string variables are not really supported. This means that you need to
add an extra dimension (with an internally assigned name) that is of
the length of the longest string in the array. But when you are
creating the string variable, the Numeric interface provides no way
of knowing this longest strlen value. Numpy does --  via the Sxx
type specifier, but it would require extensive amounts of new code to
implement this. (I should have thought of this before I added a string
array variable to the ensemble GRIB file representation. However, this
will need to be handled for the NetCDF 4 / HDF 5 implementation
anyway.