Mods to get two separate packages built for Numeric and numpy.
If no environment variable is set, then a Numeric version of PyNIO
will be built and installed to the package PyNIO.

If USE_NUMPY is set then a numpy version of PyNIO will be built
and installed to the package PyNIO.

If USE_NUMERPY is set (in which case USE_NUMPY will be ignored),
then both a Numeric and numpy version of PyNIO will be built, and
installed to packages PyNIO and PyNIO_numpy respectively.

For now, I will be setting USE_NUMERPY. The other two cases are for
when we can safely move completely over to numpy.

I've made similar changes to PyNGL's setup.py, which I've already
checked in.