This is the first check-in of PyNIO with support for coordinate selection and masked arrays. It uses a delegation pattern to achieve an effect similar to sub-classing. The types exported by the C API are now called _NioFile and _NioVariable and the module is called _Nio. Additional capabilities are added using Python code in the Nio module. In particular it imports the coordsel.py module which implements the coordinate selection features. In turn coordsel.py imports xarray.py which implements some lower level functionality used for coordinate selection.
Thanks to Juerg Schmidli for the coordinate selection code.

Some new tests are provided. Juerg provided one of them which tests the
coordinate selection code. This test is interesting because it uses the
numpy testing harness, a good idea in my opionion. The other tests the
masked array capability, and is modelled on (borrows heavily from) Juerg's
test.