Fixed some problems with dimension matching in the C API. A core dump was occurring (rather than an exception) when a numpy array was assigned to a file variable that contained fewer elements than required. Now dimensions must match more rigorously. Single element dimensions on the left side ('slow' low-numbered) of either a file variable or the assigned numpy array are discounted.
Also fixed the way the fill value is assigned to ensure that it keeps the
required numpy type. This is for when a masked array is assigned to an NioVariable.

Rewrote ma_test.py to make sure it compares the fill_value using numpy values
of the same type. Also modified the way the 'PT' variable is created: it no
longer uses random numbers, but instead uses numpy.arange. This is to ensure
repeatability of the results.