This check-in adds Juerg's fixes for problems with the auxiliary 2-D coordinate subscripting.
Also a simple fix to allow the syntax for the coordinate subscripting to contain
1 or more of any whitespace characters between dimensional elements of the subscripting
specification string.

Also fixes to allow the tests that use the Numpy test harness to continue to work.
(Note this is most likely a temporary fix because the NumPy test harness is being
revamped to use another product called 'nose'. Therefore it's current implementation
is deprecated.)

Also un-tabified all the Python module code to conform better to NumPy/Python
conventions. I considered doing it for the C code as well but decided against it
because it changes almost every line and would make subsequent diffs more
difficult.