Nio.py: added a 'MaskedExplicit' masking mode. When set enables 'ExplicitFillValues' a single value or a sequence, allowing multiple discreet values to be used for masking. The first one on the list becomes *the* fill_value. Also enabled are options 'MaskBelowValue' and 'MaskAboveValue' which allow extreme values to be masked, or, if 'MaskBelowValue' is greater than 'MaskAboveValue', specify a band of values to be masked.
Also created a 'set_option' method that allows particular options to be set on an already open
file. For now, only the masking options can be set this way.

Also added code to allow a new optional parameter to open_file called 'format'. This is a
string that can be used instead of adding an extension to a filename that allows the format
to be specified. All the recognized format extensions (without the initial '.') can be used
as indicators of the format.

coordsel.py: added some error handling.

niomodule.c: implemented the new optional 'format' keyword parameter at the C level.

Updated tests to handle new code.

CVS