Fixed a problem with NioVariable_WriteArray that the compiler warned about.
Modified setup.py to use the include files and libnio.a from the ./libsrc
directory.

This is a transitional stage.
There are still a couple of dependencies on NCARG_ROOT
1) The GRIB2 tables are copied from the grib2_codetables directory.
2) The library is still built using a yMakefile and thus depends on
the NCARG config directory.

But the library source itself has all dependencies on the HLUs or NCARG
removed.
For now to build:
make install in the NCL directory to copy all the files shared by NCL and NIO
to the pynio/libsrc directory.
Then cd into thelibsrc directory and run ymake; make includes depend; make
Then cd .. ; python setup.py install