Addresses PYNIO-48, PYNIO-46, PYNIO-50
Deleted the now unnecessary xarray.py (replaced by _xarray.py).
Modified setup.py to avoid an exception if HAS_GRIB2 is not set to 1.
The data_files variable was being referenced before initialization.
Now it is initialized to an empty list if HAS_GRIB2 is < 1.
Also modified the treatment of HAS_SZLIB. Since it is built by default
as part of HDF5, the default is to set it to 1 if HAS_NETCDF4 is set 
greater than 0.
Added the nio*.py examples to the examples directory (since it already
existed and contained the example nio05.py).