Fix for PYNIO-60.
Modifying the "language =" from "C" to "c++" causes the system's c++ compiler
to be used instead of the fortran compiler. This is necessary to prevent 
undefined references to C++ library routines on some systems that result from
some NetCDF/OPeNDAP and/or GDAL c++ code. The change works because of the
particular logic contained in numpy.distutils