PYNIO-67: coordsel bug reported by Juerg Schmidli
This could be considered a bit controversial but I am
deferring to the original author on this: with this change
if you use an extended selection that has coordinate names
that don't get matched, they are just ignored. This allows
you to use the same extended selection string for a variable and
all its coordinates -- which is definitely useful. If none of the
coordinate names match the whole variable will be returned, as if you
had asked for var[:].