PYNIO-15: problem writing netCDF files under ipython environment.
This fixes a serious memory problem. I don't think it has anything to do with ipython or netCDF 
specifically. Basically the problem was that the NIO library was "stealing" memory when attributes
with numerical values were being set. The niomodule keeps a copy of all attribute values in a Python
dictionary. When files were closed the data was being freed by the Nio library and therefore the
dictionary values were corrupted. This only showed up because the user code was keeping a reference
to the attributes after the file was closed.