PYNIO-72,PYNIO-70: Fix for Nio built without GRIB2 support and fix for version checking code in Nio.py
ma_test.py modified to handle a masked array interface change: the method get_fill_value()
and the attribute _fill_value now return the fill value as a NumPy array scalar (I think
is the term) rather than as a atomic type like float. I have modified the code to handle 
either way, so it should work with older versions of numpy as well as newer ones.

setup.py now creates a dictionary with the optional formats as keys and 1 or 0 as the value depending
on whether they are enabled in the PyNIO version. This dictionary is used in Nio.py to check whether
GRIB2 is enabled. If not it does not need to find the grib2_codetables directory. It is also used
in the test file and opendap reading code to check whether particular formats are enabled prior to
trying to read files of that format.

The version checking code has the user-suggested fix to handle NumPy release candidate versions that 
have letters after the minor release number.