PYNIO-69: adding ability to print and traverse groups
I added a 'gnode' member to the NioFile class so that when it is used 
for groups you can access the correct NclFileGrpNode.
Also modified the code so that the groups get recorded correctly.
However, there is still a problem with the variable dictionary -- it does not seem to account for
groups at all yet and variables with the same name in different groups seem to get lost.