PYNIO-69: improved support for groups.
Added support for both a hierarchical and a single top level dictionary
for groups and variables. Attributes and dimensions are also handled this 
way although it is not yet fully implemented. Handling of advanced variable
types like vlen and compound is partially implemented.
All tests in examples and test should run.