PYNIO-69: resolved issues with reference counting in the C API code
Now files are automatically closed when the Python file variable is deleted,
and conversely the data structures are mostly torn down when the close method
is called on the file. In the close case, the variable remains but the only
remaining callable method is "repr", which will state the file name and that
its mode is "closed". Attempts to call any other method or get an attribute 
generate an exception. 
I have eliminated all the memory leakage I could find from the Python part of PyNIO.
Wei is looking into a few remaining leaks from the NCL code.